Adds support for Empty House map
Adds the Empty House map as a new target location.

This includes defining the map in the target_map enum,
retrieving its English name, and specifying the minimap
and monster regions for accurate game interaction.

Also includes images for the map's monsters and the lines used for up/down calculations.